touchpad: ignore motion on finger-up
It's unlikely to be possible for human to consciously change finger
position and lift a finger so fast, that they fall into the same event.
It's a spurious motion, ignore it.

Real world significance: subjectively scrolling/cursor positioning with
a touchpad now a bit better on SAMSUNG NP305V5A laptop.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>